requirements.txt
# Core
pytest==8.3.3
pytest-asyncio==0.23.8
pytest-xdist==3.6.1
pytest-rerunfailures==14.0
allure-pytest==2.13.5

# Web/MWeb
playwright==1.47.0
pytest-playwright==0.5.0

# Mobile Native
Appium-Python-Client==4.0.0

# API
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.4.0
PyYAML==6.0.1

# Performance
locust==2.31.6

# Quality
black==24.8.0
ruff==0.5.7
mypy==1.11.1
python-dotenv==1.0.1
